CLN: address TODOs/FIXMEs
