+++ pandas/tests/indexes/ranges/test_range.py
@@ -204,10 +204,10 @@ class TestRangeIndex(NumericBase):
         loc = [1, 2, 3, 4]
         result = idx.delete(loc)
         expected = RangeIndex(0, 6, 5)
-        tm.assert_index_equal(result, expected, exact="equiv")  # TODO: retain!
+        tm.assert_index_equal(result, expected, exact=True)
 
         result = idx.delete(loc[::-1])
-        tm.assert_index_equal(result, expected, exact="equiv")  # TODO: retain!
+        tm.assert_index_equal(result, expected, exact=True)
 
     def test_delete_all_preserves_rangeindex(self):
         idx = RangeIndex(0, 6, 1)
