+++ pandas/tests/scalar/period/test_asfreq.py
@@ -428,9 +428,6 @@ class TestFreqConversion:
         ival_D_saturday = Period(freq="D", year=2007, month=1, day=6)
         ival_D_sunday = Period(freq="D", year=2007, month=1, day=7)
 
-        # TODO: unused?
-        # ival_D_monday = Period(freq='D', year=2007, month=1, day=8)
-
         ival_B_friday = Period(freq="B", year=2007, month=1, day=5)
         ival_B_monday = Period(freq="B", year=2007, month=1, day=8)
