+++ pandas/tests/series/methods/test_asof.py
@@ -2,8 +2,10 @@ import numpy as np
 import pytest
 
 from pandas._libs.tslibs import IncompatibleFrequency
+from pandas.compat import np_datetime64_compat
 
 from pandas import (
+    DatetimeIndex,
     Series,
     Timestamp,
     date_range,
@@ -15,6 +17,20 @@ import pandas._testing as tm
 
 
 class TestSeriesAsof:
+    def test_asof_nanosecond_index_access(self):
+        ts = Timestamp("20130101").value
+        dti = DatetimeIndex([ts + 50 + i for i in range(100)])
+        ser = Series(np.random.randn(100), index=dti)
+
+        first_value = ser.asof(ser.index[0])
+
+        # this used to not work bc parsing was done by dateutil that didn't
+        #  handle nanoseconds
+        assert first_value == ser["2013-01-01 00:00:00.000000050+0000"]
+
+        expected_ts = np_datetime64_compat("2013-01-01 00:00:00.000000050+0000", "ns")
+        assert first_value == ser[Timestamp(expected_ts)]
+
     def test_basic(self):
 
         # array or list or dates
